requirements.txt
# --------- pytorch --------- #
torch==1.8.1
torchvision==0.9.1
pytorch-lightning>=1.4.3
torchmetrics>=0.3.2

# --------- pytorch ecosystem --------- #
einops>=0.3.0
timm>=0.4.9
# --extra-index-url https://developer.download.nvidia.com/compute/redist
# nvidia-dali-cuda100

# --------- hydra --------- #
hydra-core==1.1.0.rc1
hydra-colorlog==1.1.0.dev1
hydra-optuna-sweeper==1.1.0.dev2
# hydra-ax-sweeper==1.1.0
# hydra-ray-launcher==0.1.2
# hydra-submitit-launcher==1.1.0

# --------- loggers --------- #
wandb>=0.10.30
# neptune-client
# mlflow
comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit
black
isort
flake8

# --------- others --------- #
jupyterlab
python-dotenv
rich
pytest
sh
scikit-learn
seaborn
pudb
imageio
# dvc
